+++ src/addons/display_commander/settings/streamline_tab_settings.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
-#include <string>
 #include "../ui/new_ui/settings_wrapper.hpp"
+
+#include <vector>
 
 namespace settings {